Drop Algebra.Graph.AdjacencyIntMap.Internal
