+++ algebraic-graphs.cabal
@@ -68,7 +68,6 @@ library
     exposed-modules:    Algebra.Graph,
                         Algebra.Graph.AdjacencyIntMap,
                         Algebra.Graph.AdjacencyIntMap.Algorithm,
-                        Algebra.Graph.AdjacencyIntMap.Internal,
                         Algebra.Graph.AdjacencyMap,
                         Algebra.Graph.AdjacencyMap.Algorithm,
                         Algebra.Graph.AdjacencyMap.Internal,
