+++ src/Data/Graph/Typed.hs
@@ -22,17 +22,13 @@ module Data.Graph.Typed (
     dfsForest, dfsForestFrom, dfs, topSort
     ) where
 
-import Algebra.Graph.AdjacencyMap.Internal    as AM
-import Algebra.Graph.AdjacencyIntMap.Internal as AIM
-
 import Data.Tree
 import Data.Maybe
 
-import qualified Data.Graph         as KL
-import qualified Data.Map.Strict    as Map
-import qualified Data.IntMap.Strict as IntMap
-import qualified Data.Set           as Set
-import qualified Data.IntSet        as IntSet
+import qualified Data.Graph as KL
+
+import qualified Algebra.Graph.AdjacencyMap    as AM
+import qualified Algebra.Graph.AdjacencyIntMap as AIM
 
 -- | 'GraphKL' encapsulates King-Launchbury graphs, which are implemented in
 -- the "Data.Graph" module of the @containers@ library.
@@ -46,25 +42,25 @@ data GraphKL a = GraphKL {
     -- Returns 'Nothing' if the argument is not in the graph.
     toVertexKL :: a -> Maybe KL.Vertex }
 
--- | Build 'GraphKL' from an 'AdjacencyMap'.
--- If @fromAdjacencyMap g == h@ then the following holds:
+-- | Build 'GraphKL' from an 'AM.AdjacencyMap'. If @fromAdjacencyMap g == h@
+-- then the following holds:
 --
 -- @
--- map ('fromVertexKL' h) ('Data.Graph.vertices' $ 'toGraphKL' h)                               == 'Algebra.Graph.AdjacencyMap.vertexList' g
--- map (\\(x, y) -> ('fromVertexKL' h x, 'fromVertexKL' h y)) ('Data.Graph.edges' $ 'toGraphKL' h) == 'Algebra.Graph.AdjacencyMap.edgeList' g
+-- map ('fromVertexKL' h) ('Data.Graph.vertices' $ 'toGraphKL' h)                               == 'AM.vertexList' g
+-- map (\\(x, y) -> ('fromVertexKL' h x, 'fromVertexKL' h y)) ('Data.Graph.edges' $ 'toGraphKL' h) == 'AM.edgeList' g
 -- 'toGraphKL' (fromAdjacencyMap (1 * 2 + 3 * 1))                                == 'array' (0,2) [(0,[1]), (1,[]), (2,[0])]
 -- 'toGraphKL' (fromAdjacencyMap (1 * 2 + 2 * 1))                                == 'array' (0,1) [(0,[1]), (1,[0])]
 -- @
-fromAdjacencyMap :: Ord a => AdjacencyMap a -> GraphKL a
-fromAdjacencyMap (AM.AM m) = GraphKL
+fromAdjacencyMap :: Ord a => AM.AdjacencyMap a -> GraphKL a
+fromAdjacencyMap am = GraphKL
     { toGraphKL    = g
     , fromVertexKL = \u -> case r u of (_, v, _) -> v
     , toVertexKL   = t }
   where
-    (g, r, t) = KL.graphFromEdges [ ((), v, Set.toAscList us) | (v, us) <- Map.toAscList m ]
+    (g, r, t) = KL.graphFromEdges [ ((), x, ys) | (x, ys) <- AM.adjacencyList am ]
 
--- | Build 'GraphKL' from an 'AdjacencyIntMap'.
--- If @fromAdjacencyIntMap g == h@ then the following holds:
+-- | Build 'GraphKL' from an 'AIM.AdjacencyIntMap'. If
+-- @fromAdjacencyIntMap g == h@ then the following holds:
 --
 -- @
 -- map ('fromVertexKL' h) ('Data.Graph.vertices' $ 'toGraphKL' h)                               == 'Data.IntSet.toAscList' ('Algebra.Graph.AdjacencyIntMap.vertexIntSet' g)
@@ -72,32 +68,33 @@ fromAdjacencyMap (AM.AM m) = GraphKL
 -- 'toGraphKL' (fromAdjacencyIntMap (1 * 2 + 3 * 1))                             == 'array' (0,2) [(0,[1]), (1,[]), (2,[0])]
 -- 'toGraphKL' (fromAdjacencyIntMap (1 * 2 + 2 * 1))                             == 'array' (0,1) [(0,[1]), (1,[0])]
 -- @
-fromAdjacencyIntMap :: AdjacencyIntMap -> GraphKL Int
-fromAdjacencyIntMap (AIM.AM m) = GraphKL
+fromAdjacencyIntMap :: AIM.AdjacencyIntMap -> GraphKL Int
+fromAdjacencyIntMap aim = GraphKL
     { toGraphKL    = g
-    , fromVertexKL = \u -> case r u of (_, v, _) -> v
+    , fromVertexKL = \x -> case r x of (_, v, _) -> v
     , toVertexKL   = t }
   where
-    (g, r, t) = KL.graphFromEdges [ ((), v, IntSet.toAscList us) | (v, us) <- IntMap.toAscList m ]
+    (g, r, t) = KL.graphFromEdges [ ((), x, ys) | (x, ys) <- AIM.adjacencyList aim ]
 
 -- | Compute the /depth-first search/ forest of a graph.
 --
--- In the following we will use the helper function:
+-- In the following examples we will use the helper function:
 --
 -- @
--- (%) :: (GraphKL Int -> a) -> AM.AdjacencyMap Int -> a
--- a % g = a $ fromAdjacencyMap g
+-- (%) :: (GraphKL Int -> a) -> 'AM.AdjacencyMap' Int -> a
+-- a % g = a $ 'fromAdjacencyMap' g
 -- @
--- for greater clarity. (One could use an AdjacencyIntMap just as well)
+--
+-- for greater clarity.
 --
 -- @
--- 'Algebra.Graph.AdjacencyMap.forest' (dfsForest % 'Algebra.Graph.AdjacencyMap.edge' 1 1)           == 'AM.vertex' 1
--- 'Algebra.Graph.AdjacencyMap.forest' (dfsForest % 'Algebra.Graph.AdjacencyMap.edge' 1 2)           == 'Algebra.Graph.AdjacencyMap.edge' 1 2
--- 'Algebra.Graph.AdjacencyMap.forest' (dfsForest % 'Algebra.Graph.AdjacencyMap.edge' 2 1)           == 'AM.vertices' [1, 2]
--- 'AM.isSubgraphOf' ('Algebra.Graph.AdjacencyMap.forest' $ dfsForest % x) x == True
--- dfsForest % 'Algebra.Graph.AdjacencyMap.forest' (dfsForest % x)      == dfsForest % x
+-- 'AM.forest' (dfsForest % 'AM.edge' 1 1)           == 'AM.vertex' 1
+-- 'AM.forest' (dfsForest % 'AM.edge' 1 2)           == 'AM.edge' 1 2
+-- 'AM.forest' (dfsForest % 'AM.edge' 2 1)           == 'AM.vertices' [1, 2]
+-- 'AM.isSubgraphOf' ('AM.forest' $ dfsForest % x) x == True
+-- dfsForest % 'AM.forest' (dfsForest % x)      == dfsForest % x
 -- dfsForest % 'AM.vertices' vs                 == 'map' (\\v -> Node v []) ('Data.List.nub' $ 'Data.List.sort' vs)
--- 'Algebra.Graph.AdjacencyMap.dfsForestFrom' ('Algebra.Graph.AdjacencyMap.vertexList' x) % x        == dfsForest % x
+-- 'AM.dfsForestFrom' ('AM.vertexList' x) % x        == dfsForest % x
 -- dfsForest % (3 * (1 + 4) * (1 + 5))     == [ Node { rootLabel = 1
 --                                                   , subForest = [ Node { rootLabel = 5
 --                                                                        , subForest = [] }]}
@@ -112,15 +109,24 @@ dfsForest (GraphKL g r _) = fmap (fmap r) (KL.dff g)
 -- the given vertices in order. Note that the resulting forest does not
 -- necessarily span the whole graph, as some vertices may be unreachable.
 --
+-- In the following examples we will use the helper function:
+--
 -- @
--- 'Algebra.Graph.AdjacencyMap.forest' (dfsForestFrom [1]    % 'Algebra.Graph.AdjacencyMap.edge' 1 1)       == 'AM.vertex' 1
--- 'Algebra.Graph.AdjacencyMap.forest' (dfsForestFrom [1]    % 'Algebra.Graph.AdjacencyMap.edge' 1 2)       == 'Algebra.Graph.AdjacencyMap.edge' 1 2
--- 'Algebra.Graph.AdjacencyMap.forest' (dfsForestFrom [2]    % 'Algebra.Graph.AdjacencyMap.edge' 1 2)       == 'AM.vertex' 2
--- 'Algebra.Graph.AdjacencyMap.forest' (dfsForestFrom [3]    % 'Algebra.Graph.AdjacencyMap.edge' 1 2)       == 'AM.empty'
--- 'Algebra.Graph.AdjacencyMap.forest' (dfsForestFrom [2, 1] % 'Algebra.Graph.AdjacencyMap.edge' 1 2)       == 'Algebra.Graph.AdjacencyMap.vertices' [1, 2]
--- 'Algebra.Graph.AdjacencyMap.isSubgraphOf' ('Algebra.Graph.AdjacencyMap.forest' $ dfsForestFrom vs % x) x == True
--- dfsForestFrom ('Algebra.Graph.AdjacencyMap.vertexList' x) % x               == 'dfsForest' % x
--- dfsForestFrom vs               % 'Algebra.Graph.AdjacencyMap.vertices' vs   == 'map' (\\v -> Node v []) ('Data.List.nub' vs)
+-- (%) :: (GraphKL Int -> a) -> 'AM.AdjacencyMap' Int -> a
+-- a % g = a $ 'fromAdjacencyMap' g
+-- @
+--
+-- for greater clarity.
+--
+-- @
+-- 'AM.forest' (dfsForestFrom [1]    % 'AM.edge' 1 1)       == 'AM.vertex' 1
+-- 'AM.forest' (dfsForestFrom [1]    % 'AM.edge' 1 2)       == 'AM.edge' 1 2
+-- 'AM.forest' (dfsForestFrom [2]    % 'AM.edge' 1 2)       == 'AM.vertex' 2
+-- 'AM.forest' (dfsForestFrom [3]    % 'AM.edge' 1 2)       == 'AM.empty'
+-- 'AM.forest' (dfsForestFrom [2, 1] % 'AM.edge' 1 2)       == 'AM.vertices' [1, 2]
+-- 'AM.isSubgraphOf' ('AM.forest' $ dfsForestFrom vs % x) x == True
+-- dfsForestFrom ('AM.vertexList' x) % x               == 'dfsForest' % x
+-- dfsForestFrom vs               % 'AM.vertices' vs   == 'map' (\\v -> Node v []) ('Data.List.nub' vs)
 -- dfsForestFrom []               % x             == []
 -- dfsForestFrom [1, 4] % (3 * (1 + 4) * (1 + 5)) == [ Node { rootLabel = 1
 --                                                          , subForest = [ Node { rootLabel = 5
@@ -131,26 +137,43 @@ dfsForest (GraphKL g r _) = fmap (fmap r) (KL.dff g)
 dfsForestFrom :: [a] -> GraphKL a -> Forest a
 dfsForestFrom vs (GraphKL g r t) = fmap (fmap r) (KL.dfs g (mapMaybe t vs))
 
--- | Compute the list of vertices visited by the /depth-first search/ in a graph,
--- when searching from each of the given vertices in order.
+-- | Compute the list of vertices visited by the /depth-first search/ in a
+-- graph, when searching from each of the given vertices in order.
+--
+-- In the following examples we will use the helper function:
 --
 -- @
--- dfs [1]   % 'Algebra.Graph.AdjacencyMap.edge' 1 1                 == [1]
--- dfs [1]   % 'Algebra.Graph.AdjacencyMap.edge' 1 2                 == [1,2]
--- dfs [2]   % 'Algebra.Graph.AdjacencyMap.edge' 1 2                 == [2]
--- dfs [3]   % 'Algebra.Graph.AdjacencyMap.edge' 1 2                 == []
--- dfs [1,2] % 'Algebra.Graph.AdjacencyMap.edge' 1 2                 == [1,2]
--- dfs [2,1] % 'Algebra.Graph.AdjacencyMap.edge' 1 2                 == [2,1]
+-- (%) :: (GraphKL Int -> a) -> 'AM.AdjacencyMap' Int -> a
+-- a % g = a $ 'fromAdjacencyMap' g
+-- @
+--
+-- for greater clarity.
+--
+-- @
+-- dfs [1]   % 'AM.edge' 1 1                 == [1]
+-- dfs [1]   % 'AM.edge' 1 2                 == [1,2]
+-- dfs [2]   % 'AM.edge' 1 2                 == [2]
+-- dfs [3]   % 'AM.edge' 1 2                 == []
+-- dfs [1,2] % 'AM.edge' 1 2                 == [1,2]
+-- dfs [2,1] % 'AM.edge' 1 2                 == [2,1]
 -- dfs []    % x                        == []
--- dfs [1,4] % (3 * (1 + 4) * (1 + 5))  == [1, 5, 4]
--- 'Algebra.Graph.AdjacencyMap.isSubgraphOf' ('Algebra.Graph.AdjacencyMap.vertices' $ dfs vs x) x == True
+-- dfs [1,4] % (3 * (1 + 4) * (1 + 5))  == [1,5,4]
+-- 'AM.isSubgraphOf' ('AM.vertices' $ dfs vs x) x == True
 -- @
 dfs :: [a] -> GraphKL a -> [a]
 dfs vs = concatMap flatten . dfsForestFrom vs
 
--- | Compute the /topological sort/ of a graph.
--- Unlike the (Int)AdjacencyMap algorithm this returns
+-- | Compute the /topological sort/ of a graph. Note that this function returns
 -- a result even if the graph is cyclic.
+--
+-- In the following examples we will use the helper function:
+--
+-- @
+-- (%) :: (GraphKL Int -> a) -> 'AM.AdjacencyMap' Int -> a
+-- a % g = a $ 'fromAdjacencyMap' g
+-- @
+--
+-- for greater clarity.
 --
 -- @
 -- topSort % (1 * 2 + 3 * 1) == [3,1,2]
