+++ test/Algebra/Graph/Test/AdjacencyIntMap.hs
@@ -18,6 +18,8 @@ import Algebra.Graph.Test
 import Algebra.Graph.Test.API (Mono (..), adjacencyIntMapAPI)
 import Algebra.Graph.Test.Generic
 
+import qualified Algebra.Graph.AdjacencyMap as AdjacencyMap
+
 t :: TestsuiteInt (Mono AdjacencyIntMap)
 t = ("AdjacencyIntMap.", adjacencyIntMapAPI)
 
@@ -25,6 +27,10 @@ testAdjacencyIntMap :: IO ()
 testAdjacencyIntMap = do
     putStrLn "\n============ AdjacencyIntMap ============"
     test "Axioms of graphs" (axioms :: GraphTestsuite AdjacencyIntMap)
+
+    putStrLn $ "\n============ AdjacencyIntMap.fromAdjacencyMap ============"
+    test "fromAdjacencyMap == stars . AdjacencyMap.adjacencyList" $ \x ->
+          fromAdjacencyMap x == (stars . AdjacencyMap.adjacencyList) x
 
     testConsistent           t
     testShow                 t
